[develop] refactor presentation package structure and create dialog fragment
