+++ app/src/main/java/com/example/calculator/feature/presentation/component/dialog/CustomPopUpDialogFragment.kt
@@ -1,0 +1,6 @@
+package com.example.calculator.feature.presentation.component.dialog
+
+import androidx.fragment.app.DialogFragment
+
+class CustomPopUpDialogFragment: DialogFragment() {
+}
