+++ app/src/main/java/com/example/calculator/feature/presentation/ui/list/viewmodel/ListViewModel.kt
@@ -1,11 +1,10 @@
-package com.example.calculator.feature.presentation.list.viewmodel
+package com.example.calculator.feature.presentation.ui.list.viewmodel
 
 import androidx.lifecycle.*
 import com.example.calculator.feature.domain.model.CalculationInfoItem
 import com.example.calculator.feature.domain.use_case.model.UseCase
 import com.example.calculator.feature.presentation.util.DummyData
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
