+++ app/src/main/java/com/example/calculator/feature/presentation/ui/update/UpdateFragment.kt
@@ -1,4 +1,4 @@
-package com.example.calculator.feature.presentation.update
+package com.example.calculator.feature.presentation.ui.update
 
 import android.os.Bundle
 import android.util.Log
