+++ app/src/main/res/layout/fragment_create_new_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.presentation.add.CreateNewItemFragment">
+    tools:context=".feature.presentation.ui.add.CreateNewItemFragment">
 
     <TextView
         android:id="@+id/textView2"
