+++ app/src/main/res/layout/fragment_update.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.presentation.update.UpdateFragment">
+    tools:context=".feature.presentation.ui.update.UpdateFragment">
 
     <TextView
         android:id="@+id/textView"
